export zod and trpc (#35)
It's annoying to have to install three libraries to get a CLI working.
So this PR exports `zod` (and an alias `z`) and a `trpc` instance that
is already init-ed with the right `meta` type.

Will aim to make clear that it's still totally fine to bring your own
zod or `@trpc/server`.

This is to avoid being _less_ convenient than alternatives like citty
clipanion etc. for simple use cases.

Tradeoff of doing this:

- 👍 smoother on ramp
- 👍 easier to install and uninstall 
- 👍 less typing
- 👍 more copy-pastable code
- 👎 slightly bigger package size

But since this is a server side lib a marginally bigger package is fine.
And users would need zod and trpc anyway - though of course this could
mean they have two copies in node_modules. Worth it I think

---------

Co-authored-by: Misha Kaletsky <mmkal@users.noreply.github.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>