standard-schema error (#87)
Copy/paste and minor modification of my own implementation on an issue
in standard-schema:
https://github.com/standard-schema/standard-schema/issues/84#issuecomment-2884500399

which is a copy/paste and minor modification of zod v4's prettifyError
which _almost_ works as is for any standard-schema error.

I thought about making a library but since standard-schema is so
copy/paste friendly and averse to changes of any kind, and also because
formatting is pretty subjective, I think this is better.

---------

Co-authored-by: Misha Kaletsky <mmkal@users.noreply.github.com>