[breaking change] switch to esm + test on node 18 (#152)
This (finally) switches trpc-cli to ESM, some users are getting problems
running with bun, which apparently uses node 18 somehow under the hood?
Should probably add bun/better-t-stack testing to CI but basically this
is the problem and existing workaround:
https://github.com/AmanVarshney01/create-better-t-stack/issues/598#issuecomment-3343467908
https://github.com/AmanVarshney01/create-better-t-stack/issues/638

This does the slightly unconventional thing of importing the optional
peer dependencies at the top level with a `.catch(String)` so that
otherwise synchronous functions don't have to become async just to load
zod/effect/valibot/whatever. I hope there aren't other messy
consequences of this...

as the charlie-creates bot notes, this is not a great thing to do, it'll
slow down startup time a little. it's a bigger breaking change to make
the various `parseXYZ` procedure async though, so I'll save that for a
later PR

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>