[experimental] proxify helper to create a cli for a remote server (#153)
allow "proxifying" a router (for want of a better word for this)

takes an existing router, but instead of calling its procedures
directly, a clone is made, where the implementation uses a trpc client
to hit a running server instead. You can create the client however you
want, but usually it'll be via `createClient` like:

```ts
proxify(router, async () => {
  return createTRPCClient<typeof router>({
    links: [httpLink({url: 'http://localhost:7500'})],
  })
})
```

e.g. in proxy.test.ts, with a server running on port 7500:

```ts
test('proxy', async () => {
  const proxiedRouter = proxify(router, async () => {
    return createTRPCClient<typeof router>({
      links: [httpLink({url: 'http://localhost:7500'})],
    })
  })
  // procedures and input types match that of `router`, but the "implementation" of each procedure is to hit the remote server over http
  expect(await run(proxiedRouter, ['greeting', '--name', 'Bob'])).toMatchInlineSnapshot(`"Hello Bob"`)
  expect(await run(proxiedRouter, ['deeply', 'nested', 'farewell', '--name', 'Bob'])).toMatchInlineSnapshot(
    `"Goodbye Bob"`,
  )
})
```